exported ValidMsgIncrementBlockRewardsPool
